@@ -3,14 +3,16 @@ using System;
 using Bejebeje.DataAccess.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
 namespace Bejebeje.DataAccess.Migrations
 {
   [DbContext(typeof(BbContext))]
-  partial class BbContextModelSnapshot : ModelSnapshot
+  [Migration("20190928222224_InitialCreate")]
+  partial class InitialCreate
   {
-    protected override void BuildModel(ModelBuilder modelBuilder)
+    protected override void BuildTargetModel(ModelBuilder modelBuilder)
     {
 #pragma warning disable 612, 618
       modelBuilder
